+++ src/utils/addGear.ts
@@ -1,4 +1,4 @@
-import type { Gear } from '../../types';
+import type { Gear } from '../types';
 import { getGearCollection } from './database';
 
 export async function addGear(gear: Gear): Promise<void> {
